现在只差指令解析，电磁锁，

+++ 00STM32C8/APP_F103C8/BSP/bsp_dht22.h
@@ -16,7 +16,7 @@
 														
 			
 
-u8 DHT22_Read_Data(u8 *humi,u8 *temp);	//DHT22读取温度
+u8 DHT22_Read_Data(u16 *humi,u16 *temp);	//DHT22读取温度
 
 #endif
